Refactor benchmark logging and configuration; update sweep value type to uint64_t, enhance log output format, and improve memory allocation handling in radix sort benchmark

+++ benchmarks/include/benchmark.h
@@ -34,7 +34,7 @@ extern "C" {
         const char *logs_dir;
         MPI_File *mpi_log_file;
 
-        unsigned int sweep_value;
+        uint64_t sweep_value;
         char sweep_name[64];
     } BenchmarkConfig;
